backend/requirements.txt
# FastAPI y dependencias
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Base de datos
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AWS
boto3==1.34.14
botocore==1.34.14

# Utilidades
python-dateutil==2.8.2
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
httpx==0.26.0

# Manejo de archivos
python-magic==0.4.27
aiofiles==23.2.1

# Logging y monitoreo
loguru==0.7.2
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Desarrollo
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
